Inserted the Location endpoint

+++ Motorcycle-Rental-Application/UseCases/MotorcycleUseCase/DeleteMotorcycleUseCase.cs
@@ -7,11 +7,11 @@ using Motorcycle_Rental_Infrastructure.Interfaces;
 
 namespace Motorcycle_Rental_Application.UseCases.Motorcycle
 {
-    public class DeleteMotorcycleUseCase(IMotorcycleRepository motorcycleRepository, ILogger<DeleteMotorcycleUseCase> logger) : IDeleteMotorcycleUseCase
+    public class DeleteMotorcycleUseCase(IMotorcycleRepository motorcycleRepository, ILogger<DeleteMotorcycleUseCase> logger, ILocationRepository locationRepository) : IDeleteMotorcycleUseCase
     {
         private readonly ILogger _logger = logger;
         private readonly IMotorcycleRepository _motorcycleRepository = motorcycleRepository;
- 
+        private readonly ILocationRepository _locationRepository = locationRepository;
 
         public async Task<Result> ExecuteAsync(DeleteMotorcycleDTO request)
         {
@@ -34,6 +34,12 @@ namespace Motorcycle_Rental_Application.UseCases.Motorcycle
                 return Result.Fail("Motorcycle not found");
             }
 
+            var locationVerify = await _locationRepository.RecoverByAsync(l=>l.Motorcycle_Id.Equals(motorcycleResult.Identifier));
+            
+            if (locationVerify is not null)
+            {
+                return Result.Fail("The motorcycle is currently leased and cannot be deleted.");
+            }
             await _motorcycleRepository.DeleteAsync(motorcycleResult);
